add stream & search
