+++ Backend/config.py
@@ -1,0 +1,5 @@
+# Enter Twitter API Keys
+access_token = "1254430088793223168-2vNjv4WYXrfkvAnwxpmYvXDeJiL2SI"
+access_token_secret = "NjGEh6pK6oUru4WwndDzDdrIgms7zn5M8LY52Ryg1HQGK"
+consumer_key = "be5MWgU3xxhxL3yXBW5P84vtL"
+consumer_secret = "7KLLsiXLsR71arTnBwQNc14EDaE7UZrpsepINb4Q8UauLOjWb8"
