+++ Backend/utils.py
@@ -3,7 +3,6 @@
 # Last-update: 9/5/2020
 
 import couchdb
-from tweepy import StreamListener
 
 
 # Connect CouchDB
@@ -18,14 +17,3 @@ class DatabaseConnection:
 
     def get_db(self):
         return self.server[self.db_name]
-
-
-# Create the class that will handle the tweet stream
-class StdOutListener(StreamListener):
-
-    def on_data(self, data):
-        print(data)
-        return True
-
-    def on_error(self, status):
-        print(status)
